requirements.txt
discord.py
pyspellchecker
aiohttp
